@@ -5,14 +5,15 @@ homepage: https://tomas-chyly.com/en/project/jstrions/
 repository: https://github.com/tomaschyly/JsTrions
 issue_tracker: https://github.com/tomaschyly/JsTrions/issues
 
-publish_to: 'none'
+publish_to: "none"
 
-version: 1.5.1+17 #Win version needs to change also in Runner.rc & msix
+version: 1.6.0+26
+#Win version needs to change also in Runner.rc & msix
 #Snap needs to update version too in snapcraft.yaml
 
 environment:
-  sdk: ">=2.12.0-0 <3.0.0"
-  flutter: ">=2.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.0.0"
 
 dependencies:
   flutter:
@@ -20,31 +21,35 @@ dependencies:
   flutter_localizations:
     sdk: flutter
 
-  tch_appliable_core: ^0.25.0
-  tch_common_widgets: ^0.27.0
+  tch_appliable_core: ^0.35.0
+  tch_common_widgets: ^0.38.0
   
   supercharged: ^2.1.1
-  package_info_plus: ^4.1.0
-  window_manager: ^0.3.6
-  archive: ^3.4.3
+  package_info_plus: ^9.0.0
+  window_manager: ^0.5.1
+  archive: ^4.0.7
 
-  cupertino_icons: ^1.0.6
-  flutter_svg: ^2.0.7
+  cupertino_icons: ^1.0.8
+  flutter_svg: ^2.2.1
 
-  intl: ^0.18.1
+  intl: ^0.20.2
   html_unescape: ^2.0.0
+  jiffy: ^6.4.3
+  bot_toast: ^4.1.3
 
-  path: ^1.8.2
-  path_provider: ^2.0.14
-  sqflite: ^2.3.0
-  sembast: ^3.5.0+1
+  path: ^1.9.1
+  path_provider: ^2.1.5
+  sqflite: ^2.4.2
+  sembast: ^3.8.5+1
 
-  file_selector: ^0.9.2+5
+  file_selector: ^1.0.4
 
-  url_launcher: ^6.1.10
+  url_launcher: ^6.3.2
+
+  openai_dart: ^0.5.5
 
 dev_dependencies:
-  msix: ^2.8.18
+  msix: ^3.16.12
 
 #dependency_overrides:
 #  intl: ^0.18.1
@@ -66,18 +71,24 @@ flutter:
         - asset: fonts/Kalam-Regular.ttf
 
 # https://pub.dev/packages/msix
-# flutter build windows
-# flutter pub run msix:create
+# flutter clean && flutter build windows
+# flutter clean && dart run msix:create
 msix_config:
   display_name: JsTrions
   publisher_display_name: Tomas Chyly
   publisher: CN=B0FFABE7-CCEC-4059-A43F-BC992CC5B5D2
   identity_name: 37940TomasChyly.JsTrions
-  msix_version: 1.5.1.0
+  msix_version: 1.6.0.0
   icons_background_color: transparent
   architecture: x64
   capabilities: "internetClient"
-  logo_path: "G:/Development/FlutterProjs/JsTrions/windows/runner/resources/icon.png"
-  start_menu_icon_path: "G:/Development/FlutterProjs/JsTrions/windows/runner/resources/icon.png"
-  tile_icon_path: "G:/Development/FlutterProjs/JsTrions/windows/runner/resources/icon.png"
+  logo_path: "E:/Development/FlutterProjs/JsTrions/windows/runner/resources/icon.png"
+  start_menu_icon_path: "E:/Development/FlutterProjs/JsTrions/windows/runner/resources/icon.png"
+  tile_icon_path: "E:/Development/FlutterProjs/JsTrions/windows/runner/resources/icon.png"
   store: true
+# https://docs.flutter.dev/deployment/linux
+# https://documentation.ubuntu.com/snapcraft/stable/how-to/publishing/publish-a-snap/
+# flutter clean && snapcraft clean && snapcraft pack
+# sudo snap remove js-trions
+# sudo snap install --dangerous ./js-trions_1.6.0_amd64.snap
+# snapcraft upload --release=stable ./js-trions_1.6.0_amd64.snap
