@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:js_trions/App.dart';
+import 'package:js_trions/app.dart';
 
 void main() {
   runApp(App());
